# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Maewin API Documentation",
        "description": "API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Maewin OpenApi Server"
        }
    ],
    "paths": {
        "/signup": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "operationId": "signup",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "example": "Satawat Orachunwekhin"
                                    },
                                    "email": {
                                        "type": "email",
                                        "example": "stworchwk@gmail.com"
                                    },
                                    "password": {
                                        "type": "password",
                                        "example": "123456"
                                    },
                                    "password_confirmation": {
                                        "type": "password",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sign Up Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Successfully created user!",
                                        "user": {
                                            "id": 6,
                                            "full_name": "Chonnikan",
                                            "email": "chnkkpch@gmail.com",
                                            "created_at": "2020-10-30T08:39:13.000000Z",
                                            "updated_at": "2020-10-30T08:39:13.000000Z"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "example": "stworchwk@gmail.com"
                                    },
                                    "password": {
                                        "type": "password",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "access_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9.eyJhd....",
                                        "token_type": "Bearer",
                                        "expires_at": "2020-11-14 08:46:17"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/logout": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Login Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Successfully logged out."
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/profile": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "operationId": "profile",
                "responses": {
                    "200": {
                        "description": "Call Profile Data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "status": true,
                                        "message": "",
                                        "result": {
                                            "id": 6,
                                            "email": "chnkkpch@gmail.com",
                                            "email_verified_at": null,
                                            "social_google": null,
                                            "full_name": "Chonnikan",
                                            "nationality_id": null,
                                            "prefix_phone_number_id": null,
                                            "phone_number": null,
                                            "id_card": null,
                                            "active": 1,
                                            "created_at": "2020-10-30T08:39:13.000000Z",
                                            "updated_at": "2020-10-30T08:39:13.000000Z"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/profileUpdate": {
            "put": {
                "tags": [
                    "Authentication"
                ],
                "operationId": "profile_update",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "example": "Satawat Orachunwekhin"
                                    },
                                    "nationality_id": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "prefix_phone_number_id": {
                                        "type": "integer",
                                        "example": 46
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "953685568"
                                    },
                                    "id_card": {
                                        "type": "string",
                                        "example": "1509901361170"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "access_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9.eyJhd....",
                                        "token_type": "Bearer",
                                        "expires_at": "2020-11-14 08:46:17"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/locations": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "operationId": "locations",
                "responses": {
                    "200": {
                        "description": "Call Location Data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "status": true,
                                        "message": "",
                                        "result": [
                                            {
                                                "id": 1,
                                                "location_category_id": 1,
                                                "code": "MW9-00001",
                                                "title_th": "น้ำตกแม่สะป๊อก",
                                                "title_en": "Mae Sa Prock Waterfall",
                                                "mark_down": "fsafsafsaf",
                                                "village_name": "สันกับตอง",
                                                "village_no": 5,
                                                "address": "22/1 M.8 Saraphi",
                                                "owner_full_name": "ศตวรรษ อรชุนเวคิน",
                                                "tel": "0953685568",
                                                "latitude": "18.795824565103683",
                                                "longitude": "98.9686906612549",
                                                "destination_latitude": "18.795824565103683",
                                                "destination_longitude": "98.9686906612549",
                                                "budget": 200,
                                                "time_spent": 30,
                                                "active": 1,
                                                "thumbnail_url": "http://localhost:8000/images/locations/9fcb1c0058f8c2edf2ed2678a77b57ea98e661eb.jpg",
                                                "category": [
                                                    {
                                                        "id": 1,
                                                        "name": "น้ำตก",
                                                        "icon_url": "http://localhost:8000/images/locationCategories/6c00dbffd5ebc4ccf30b0ad40e2fc3200b988c35.png"
                                                    }
                                                ],
                                                "images": [
                                                    {
                                                        "id": 1,
                                                        "location_id": 1,
                                                        "title_th": "น้ำตกแม่สะป๊อก",
                                                        "title_en": "Mae Sa Prock Waterfall",
                                                        "detail_th": null,
                                                        "detail_en": null,
                                                        "image_url": "http://localhost:8000/images/locations/e75b99f7801c7cad86821a81b5f9c4510559d162.jpg"
                                                    }
                                                ]
                                            },
                                            {
                                                "id": 1,
                                                "location_category_id": 1,
                                                "code": "MW9-00001",
                                                "title_th": "น้ำตกแม่สะป๊อก",
                                                "title_en": "Mae Sa Prock Waterfall",
                                                "mark_down": "fsafsafsaf",
                                                "village_name": "สันกับตอง",
                                                "village_no": 5,
                                                "address": "22/1 M.8 Saraphi",
                                                "owner_full_name": "ศตวรรษ อรชุนเวคิน",
                                                "tel": "0953685568",
                                                "latitude": "18.795824565103683",
                                                "longitude": "98.9686906612549",
                                                "destination_latitude": "18.795824565103683",
                                                "destination_longitude": "98.9686906612549",
                                                "budget": 200,
                                                "time_spent": 30,
                                                "active": 1,
                                                "thumbnail_url": "http://localhost:8000/images/locations/9fcb1c0058f8c2edf2ed2678a77b57ea98e661eb.jpg",
                                                "category": [
                                                    {
                                                        "id": 1,
                                                        "name": "น้ำตก",
                                                        "icon_url": "http://localhost:8000/images/locationCategories/6c00dbffd5ebc4ccf30b0ad40e2fc3200b988c35.png"
                                                    }
                                                ],
                                                "images": [
                                                    {
                                                        "id": 1,
                                                        "location_id": 1,
                                                        "title_th": "น้ำตกแม่สะป๊อก",
                                                        "title_en": "Mae Sa Prock Waterfall",
                                                        "detail_th": null,
                                                        "detail_en": null,
                                                        "image_url": "http://localhost:8000/images/locations/e75b99f7801c7cad86821a81b5f9c4510559d162.jpg"
                                                    }
                                                ]
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/locationCategories": {
            "get": {
                "tags": [
                    "Location Categories"
                ],
                "operationId": "locationCategories",
                "responses": {
                    "200": {
                        "description": "Call Location Category Data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "status": true,
                                        "message": "",
                                        "result": [
                                            {
                                                "id": 1,
                                                "name": "น้ำตก",
                                                "icon_url": "http://localhost:8000/images/locationCategories/6c00dbffd5ebc4ccf30b0ad40e2fc3200b988c35.png"
                                            },
                                            {
                                                "id": 1,
                                                "name": "น้ำตก",
                                                "icon_url": "http://localhost:8000/images/locationCategories/6c00dbffd5ebc4ccf30b0ad40e2fc3200b988c35.png"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/nationalities": {
            "get": {
                "tags": [
                    "Nationality"
                ],
                "operationId": "nationality",
                "responses": {
                    "200": {
                        "description": "Call Nationality Data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "status": true,
                                        "message": "",
                                        "result": [
                                            {
                                                "id": 6,
                                                "name_th": "ไทย",
                                                "name_en": "Thai"
                                            },
                                            {
                                                "id": 7,
                                                "name_th": "อินเดีย",
                                                "name_en": "Indian"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/prefixPhoneNumbers": {
            "get": {
                "tags": [
                    "Prefix Phone Number"
                ],
                "operationId": "prefixPhoneNumbers",
                "responses": {
                    "200": {
                        "description": "Call Prefix Phone Number Data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "status": true,
                                        "message": "",
                                        "result": [
                                            {
                                                "id": 6,
                                                "prefix": "66",
                                                "name_th": "ไทย",
                                                "name_en": "Thai"
                                            },
                                            {
                                                "id": 7,
                                                "prefix": "91",
                                                "name_th": "อินเดีย",
                                                "name_en": "Indian"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}